docs(native components): moved all native component stories to a separate folder
