--- tegel/src/components/footer/readme.md
@@ -1,16 +1,0 @@
-# c-footer
-
-<!-- Auto Generated Below -->
-
-## Properties
-
-| Property      | Attribute      | Description                                                            | Type     | Default                     |
-| ------------- | -------------- | ---------------------------------------------------------------------- | -------- | --------------------------- |
-| `items`       | `items`        | Set footer links                                                       | `any`    | `[]`                        |
-| `socialItems` | `social-items` | Add social media icons                                                 | `any`    | `[]`                        |
-| `text`        | `text`         | Change default copyright text                                          | `string` | `'Copyright © Scania 2019'` |
-| `theme`       | `theme`        | Per default, this will inherit the value from sdds-theme name property | `string` | `undefined`                 |
-
----
-
-_Built with [StencilJS](https://stenciljs.com/)_
