--- tegel/src/components/header/readme.md
@@ -1,18 +1,0 @@
-# sdds-header
-
-
-
-<!-- Auto Generated Below -->
-
-
-## Properties
-
-| Property   | Attribute   | Description                            | Type     | Default         |
-| ---------- | ----------- | -------------------------------------- | -------- | --------------- |
-| `iconHref` | `icon-href` | URL for the sdds-header-icon           | `string` | `'#'`           |
-| `siteName` | `site-name` | The na that is displayed in the header | `string` | `'Application'` |
-
-
-----------------------------------------------
-
-*Built with [StencilJS](https://stenciljs.com/)*
